added an API to access `implementation_search_path`
