+++ gpt4all-bindings/rust/gpt4all-sys/wrapper.h
